playground/fastapi/fastapi-workbook-2/requirements.txt
# Core FastAPI
fastapi==0.115.0
uvicorn[standard]==0.32.0

# Database
sqlalchemy==2.0.35
alembic==1.13.3
psycopg2-binary==2.9.10  # PostgreSQL adapter (optional, for production)

# Authentication
python-jose[cryptography]==3.3.0  # JWT tokens
passlib[bcrypt]==1.7.4  # Password hashing
python-multipart==0.0.12  # For file uploads and form data

# File handling
python-magic==0.4.27  # File type detection
Pillow==10.4.0  # Image processing

# Configuration
pydantic-settings==2.5.2
python-dotenv==1.0.1

# Testing
pytest==8.3.3
pytest-asyncio==0.24.0
httpx==0.27.2

# Development
black==24.10.0  # Code formatting
ruff==0.7.4  # Linting
